Revert "move format functions that are only used in view-data into view-data"
This reverts commit 2263a98fd97b34ee0cf443ce18d1e21d17ede0d6.
