+++ containers/ecr-viewer/tests/unit/app/view-data/services/evaluateFhirDataServices.test.tsx
@@ -7,6 +7,7 @@ import * as _BundleWithDeceasedPatient from "../../../../../../../test-data/fhir
 import BundlePatientMultiple from "../../../../../../../test-data/fhir/BundlePatientMultiple.json";
 import * as _BundlePatientWithCovid from "../../../../../../../test-data/fhir/BundlePatientWithCovid.json";
 import BundlePractitionerRole from "../../../../../../../test-data/fhir/BundlePractitionerRole.json";
+import { formatAge } from "@/app/services/formatService";
 import { evaluateValue } from "@/app/utils/evaluate";
 import mappings from "@/app/utils/evaluate/fhir-paths";
 import {
@@ -31,7 +32,6 @@ import {
   evaluateProviderData,
   evaluatePregnancyData,
 } from "@/app/view-data/services/evaluateFhirDataService";
-import { formatAge } from "@/app/view-data/services/formatService";
 
 const BundleWithPatient = _BundleWithPatient as Bundle;
 const BundleWithDeceasedPatient = _BundleWithDeceasedPatient as Bundle;
